added post exceptions

+++ src/post/post.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import {PostDto, CreatePostDto } from './dto';
@@ -16,7 +16,23 @@ export class PostService {
     }
 
     async read(id:string) {
-        return await this.postRepository.findOne({where:{id}});
+        const post =  await this.postRepository.findOne({where:{id}});
+
+        if(!post) {
+            // 1. optioon
+            // throw new HttpException({
+            //     status: HttpStatus.NOT_FOUND,
+            //     error: 'Not Found',
+            //   }, HttpStatus.NOT_FOUND);
+            
+            // 2. option
+            throw new NotFoundException()
+           
+             // 3. option
+            throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+        }
+
+        return post;
     }
 
     async create(data:CreatePostDto) {
@@ -27,14 +43,28 @@ export class PostService {
     }
 
     async update(id:string, data:Partial<PostDto>) {
+
+        const post = await this.postRepository.findOne({where:{ id }});
+     
+        if(!post) {
+         
+            //throw new HttpException('Not Found', HttpStatus.NOT_FOUND)
+            throw new NotFoundException();
+        }
         await this.postRepository.update({id}, data);
-        return await this.postRepository.findOne({id});
+        return post;
     }
 
     async destroy(id:string) {
-        await this.postRepository.delete({id});
 
-        return { deleted:true}
+        const post = await this.postRepository.findOne({where:{ id }});
+     
+        if(!post) {
+            throw new NotFoundException();
+        }
+
+        await this.postRepository.delete({id});
+        return post;
     }
